Calibrate confidence intervals

Calibrate confidence intervals

calibrateConfidenceInterval(logRr, seLogRr, model, ciWidth = 0.95)

Arguments

logRr

A numeric vector of effect estimates on the log scale.

seLogRr

The standard error of the log of the effect estimates. Hint: often the standard error = (log(<lower bound 95 percent confidence interval>) - log(<effect estimate>))/qnorm(0.025).

model

An object of type systematicErrorModel as created by the fitSystematicErrorModel function.

ciWidth

The width of the confidence interval. Typically this would be .95, for the 95 percent confidence interval.

Value

A data frame with calibrated confidence intervals and point estimates.

Details

Compute calibrated confidence intervals based on a model of the systematic error.

Examples

data <- simulateControls(n = 50 * 3, mean = 0.25, sd = 0.25, trueLogRr = log(c(1, 2, 4)))
model <- fitSystematicErrorModel(data$logRr, data$seLogRr, data$trueLogRr)
newData <- simulateControls(n = 15, mean = 0.25, sd = 0.25, trueLogRr = log(c(1, 2, 4)))
result <- calibrateConfidenceInterval(newData$logRr, newData$seLogRr, model)
result
#>         logRr    logLb95Rr logUb95Rr   seLogRr
#> 1   0.3101451 -0.296757303 0.9079005 0.3073163
#> 2   0.4253582 -0.207643628 1.0463047 0.3198907
#> 3   1.3976490  0.847115430 1.9306055 0.2764056
#> 4  -0.1632532 -0.790346373 0.4684435 0.3211258
#> 5   0.7808329  0.200747045 1.3429875 0.2913932
#> 6   1.1439241  0.500236553 1.7699178 0.3239042
#> 7   0.2659664 -0.284050933 0.8073309 0.2784188
#> 8   0.6073182 -0.001248903 1.1979027 0.3059116
#> 9   1.2957840  0.730437372 1.8434968 0.2839489
#> 10 -0.2069522 -0.822001350 0.4140435 0.3153234
#> 11  0.6701432  0.120635488 1.2016683 0.2757787
#> 12  1.4568242  0.917120789 1.9789867 0.2708891
#> 13  0.2138665 -0.429216126 0.8510188 0.3265965
#> 14  0.5343262 -0.040450254 1.0923551 0.2889863
#> 15  1.2383607  0.698569794 1.7604890 0.2709027