Calibrate the log likelihood ratio

calibrateLlr computes calibrated log likelihood ratio using the fitted null distribution

calibrateLlr(null, likelihoodApproximation, twoSided = FALSE, upper = TRUE)

Arguments

null
An object of class null created using the fitNull function or an object of class mcmcNull created using the fitMcmcNull function.

likelihoodApproximation
Either a data frame containing normal, skew-normal, or custom parametric likelihood approximations, or a list of (adaptive) grid likelihood profiles.

twoSided
Compute two-sided (TRUE) or one-sided (FALSE) p-value?

upper
If one-sided: compute p-value for upper (TRUE) or lower (FALSE) bound?

Value

The calibrated log likelihood ratio.

Examples

data(sccs)
negatives <- sccs[sccs$groundTruth == 0, ]
null <- fitNull(negatives$logRr, negatives$seLogRr)
positive <- sccs[sccs$groundTruth == 1, ]
calibrateLlr(null, positive)

#> Detected data following normal distribution

#> [1] 0