computeTraditionalP {EmpiricalCalibration}	R Documentation

Compute the (traditional) p-value

Description

computeTraditionalP computes the traditional two-sided p-value based on the log of the relative risk and the standard error of the log of the relative risk.

Usage

computeTraditionalP(logRr, seLogRr, twoSided = TRUE, upper = TRUE)

Arguments

logRr
A numeric vector of one or more effect estimates on the log scale

seLogRr
The standard error of the log of the effect estimates. Hint: often the standard error = (log(<lower bound 95 percent confidence interval>) - log(<effect estimate>))/qnorm(0.025)

twoSided
Compute two-sided (TRUE) or one-sided (FALSE) p-value?

upper
If one-sided: compute p-value for upper (TRUE) or lower (FALSE) bound?

Value

The (traditional) p-value.

Examples

data(sccs)
positive <- sccs[sccs$groundTruth == 1, ]
computeTraditionalP(positive$logRr, positive$seLogRr)

[Package EmpiricalCalibration version 3.1.0 Index]