fitNull {EmpiricalCalibration}	R Documentation

Fit the null distribution

Description

fitNull fits the null distribution to a set of negative controls

Usage

fitNull(logRr, seLogRr)

Arguments

logRr
A numeric vector of effect estimates on the log scale

seLogRr
The standard error of the log of the effect estimates. Hint: often the standard error = (log(<lower bound 95 percent confidence interval>) - log(<effect estimate>))/qnorm(0.025)

Details

This function fits a Gaussian function to the negative control estimates as described in Schuemie et al (2014).

Value

An object containing the parameters of the null distribution.

References

Schuemie MJ, Ryan PB, Dumouchel W, Suchard MA, Madigan D. Interpreting observational studies: why empirical calibration is needed to correct p-values. Statistics in Medicine 33(2):209-18,2014

Examples

data(sccs)
negatives <- sccs[sccs$groundTruth == 0, ]
null <- fitNull(negatives$logRr, negatives$seLogRr)
null

[Package EmpiricalCalibration version 2.1.0 Index]