fitNullNonNormalLl {EmpiricalCalibration}	R Documentation

Fit the null distribution using non-normal log-likelihood approximations

Description

fitNullNonNormalLl fits the null distribution to a set of negative controls

Usage

fitNullNonNormalLl(likelihoodApproximations)

Arguments

likelihoodApproximations
Either a data frame containing normal, skew-normal, or custom parametric likelihood approximations, or a list of (adaptive) grid likelihood profiles.

Details

This function fits a Gaussian function to the negative control estimates, using non-normal approximations of the per-negative control log likelihood.

Value

An object containing the parameters of the null distribution.

Examples

data(sccs)
negatives <- sccs[sccs$groundTruth == 0, ]
null <- fitNullNonNormalLl(negatives)
null

[Package EmpiricalCalibration version 3.1.0 Index]