Fit a systematic error model

Fit a systematic error model

fitSystematicErrorModel(logRr, seLogRr, trueLogRr,
  estimateCovarianceMatrix = TRUE)

Arguments

logRr
A numeric vector of effect estimates on the log scale.

seLogRr
The standard error of the log of the effect estimates. Hint: often the standard error = (log(<lower bound 95 percent confidence interval>) - log(<effect estimate>))/qnorm(0.025).

trueLogRr
A vector of the true effect sizes.

estimateCovarianceMatrix
should a covariance matrix be computed? If so, confidence intervals for the model parameters will be available.

Value

An object of type systematicErrorModel.

Details

Fit a model of the systematic error as a function of true effect size. This model is an extension of the method for fitting the null distribution. The mean and log(standard deviations) of the error distributions are assumed to be linear with respect to the true effect size, and each component is therefore represented by an intercept and a slope.

Examples

controls <- simulateControls(n = 50 * 3, mean = 0.25, sd = 0.25, trueLogRr = log(c(1, 2, 4)))
model <- fitSystematicErrorModel(controls$logRr, controls$seLogRr, controls$trueLogRr)
model

#>  meanIntercept      meanSlope logSdIntercept     logSdSlope 
#>     0.25535879     1.01729497    -1.40972250    -0.02597799 
#> attr(,"CovarianceMatrix")
#>               [,1]          [,2]          [,3]          [,4]
#> [1,]  1.181744e-03 -1.024144e-03  1.168047e-06 -3.452534e-05
#> [2,] -1.024144e-03  1.455294e-03 -3.094007e-05  1.881897e-05
#> [3,]  1.168047e-06 -3.094007e-05  1.094728e-02 -9.480165e-03
#> [4,] -3.452534e-05  1.881897e-05 -9.480165e-03  1.410706e-02
#> attr(,"LB95CI")
#> [1]  0.1879821  0.9425256 -1.6147921 -0.2587691
#> attr(,"UB95CI")
#> [1]  0.3227355  1.0920643 -1.2046529  0.2068131
#> attr(,"class")
#> [1] "systematicErrorModel"