computeExpectedAbsoluteSystematicError {EmpiricalCalibration}	R Documentation

Compute the expected absolute systematic error

Description

For a random study estimate, what is the expected value of the absolute systematic error? Provides a single summary value for a null distribution. The expected systematic error of a null distribution is equal to its mean (mu), and is insensitive to the spread of the null distribution (sigma).

Taking the absolute value of the expected systematic error we can express both mean and spread of the estimated null distribution.

Usage

computeExpectedAbsoluteSystematicError(null, alpha = 0.05)

Arguments

null
An object of class null created using the fitNull function or an object of class mcmcNull created using the fitMcmcNull function.

alpha
The expected type I error for computing the credible interval.

Value

The expected absolute systematic error. If the provided null argument is of type mcmcNull, the credible interval (defined by alpha) is also returned.

Examples

data(sccs)
negatives <- sccs[sccs$groundTruth == 0, ]
null <- fitNull(negatives$logRr, negatives$seLogRr)
computeExpectedAbsoluteSystematicError(null)

[Package EmpiricalCalibration version 2.1.0 Index]