computeTraditionalP {EmpiricalCalibration}	R Documentation

Compute the (traditional) p-value

Description

computeTraditionalP computes the traditional two-sided p-value based on the log of the relative risk and the standard error of the log of the relative risk.

Usage

computeTraditionalP(logRr, seLogRr)

Arguments

logRr
A numeric vector of one or more effect estimates on the log scale

seLogRr
The standard error of the log of the effect estimates. Hint: often the standard error = (log(<lower bound 95 percent confidence interval>) - log(<effect estimate>))/qnorm(0.025)

Value

The two-sided (traditional) p-value.

Examples

data(sccs)
positive <- sccs[sccs$groundTruth == 1, ]
computeTraditionalP(positive$logRr, positive$seLogRr)

[Package EmpiricalCalibration version 2.1.0 Index]